refactor: name the module amqp091
in case we add separate support for amqp v1.0.0 in future
